Разбил js на модули
Возможно можно вынести createPhoto и getPhoto в отдельный модуль от значений(констант), но не уверен насколько это необходимо
